ai-service/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
Pillow>=10.0.0
numpy>=1.24.0
gunicorn>=21.2.0
# Uncomment for real YOLOv8 model detection:
# ultralytics>=8.1.0
# torch>=2.0.0
